Bug 781341 - Change smoketest to not delete files, add machine metadata;r=mcote
The current smoke test assumes it will only be run once, and thus deletes
the smoketest_pass and smoketest_fail files when run. To do our smoketest
across a number of machines, we want to be able to run the smoketest more
than once and aggregate results by appending to the smoketest_pass and
smoketest_fail files, so don't do this.